Added 32-bits unsigned int size_t shift operation fix to client_ws.hpp as well

+++ client_ws.hpp
@@ -155,7 +155,7 @@ namespace SimpleWeb {
                 }
                 
                 for(int c=num_bytes-1;c>=0;c--) {
-                    send_stream->put((length>>(8*c))%256);
+                    send_stream->put((static_cast<unsigned long long>(length) >> (8 * c)) % 256);
                 }
             }
             else
